Clean up and some restructering of files/code
