+++ utils/modelling/main_modeler.py
@@ -1,66 +1,15 @@
-from interpret.glassbox import ExplainableBoostingClassifier, ExplainableBoostingRegressor
-import sys
-import pandas as pd
-import numpy as np
-from matplotlib import pyplot as plt
 import pickle
 import random
 from sklearn.model_selection import train_test_split
-from utils.checks import *
 from utils.modelling.performance import *
 from utils.modelling.calibration import *
 
-def trainEBM(X_train, y_train, params, model_type, logging):
-    if 'feature_names' not in params.keys():
-        params['feature_names'] = X_train.columns
-    if model_type == 'regression':
-        clf = ExplainableBoostingRegressor(**params)
-    elif model_type == 'classification':
-        clf = ExplainableBoostingClassifier(**params)
-    else:
-        print('Only regression or classification available')
-        logging.warning('Only regression or classification available')
+# The actual algorithms (grey as we refer to them dynamically)
+from utils.modelling.models import ebm
+from utils.modelling.models import decision_rule
+from utils.modelling.models import decision_tree
 
-    clf.fit(X_train, y_train)
-    print('Trained explainable boosting machine \n')
-    logging.info('Trained explainable boosting machine')
-
-    return clf
-
-
-def featureExplanationSave(ebm, given_name, file_type, logging):
-
-    ebm_global = ebm.explain_global()
-
-    # Save overall feature importance graph
-    plotly_fig = ebm_global.visualize()
-
-    if file_type == 'png':
-        plotly_fig.write_image('{given_name}/1_overall_feature_importance.png'.format(given_name=given_name))
-    elif file_type == 'html':
-        plotly_fig.write_html('{given_name}/1_overall_feature_importance.html'.format(given_name=given_name))
-
-    # Save feature specific explanation graphs
-    for index, value in enumerate(ebm.feature_groups_):
-        plotly_fig = ebm_global.visualize(index)
-
-        # reformatting feature name
-        feature_name = ebm.feature_names[index]
-        chars = "\\`./"
-        for c in chars:
-            if c in feature_name:
-                feature_name = feature_name.replace(c, "_")
-
-        if file_type == 'png':
-            plotly_fig.write_image('{given_name}/explain_{feature}.png'.format(given_name=given_name, feature=feature_name))
-        elif file_type == 'html':
-            # or as html file
-            plotly_fig.write_html('{given_name}/explain_{feature}.html'.format(given_name=given_name, feature=feature_name))
-
-    print('Explanation plots of {n_features} features saved'.format(n_features=index+1))
-    logging.info('Explanation plots of {n_features} features saved'.format(n_features=index+1))
-
-def make_model(given_name, datasets, model_type, model_params, post_params, logging):
+def make_model(given_name, datasets, model_name, model_type, model_params, post_params, logging):
     # unpack datasets
     X_train = datasets['cv_train']['X']
     y_train = datasets['cv_train']['y']
@@ -86,8 +35,7 @@ def make_model(given_name, datasets, model_type, model_params, post_params, logg
             else:
                 X_slice_train, y_slice_train = X_train[fold_id], y_train[fold_id]
 
-            clf = trainEBM(X_slice_train, y_slice_train, model_params, model_type, logging)
-            logging.info(f'Model params:\n {clf.get_params}')
+            clf = globals()[model_name].trainModel(X_slice_train, y_slice_train, model_params, model_type, logging)
 
             if post_params['calibration'] != 'false':
                 clf = calibrateModel(clf, X_cal, y_cal, logging, method=post_params['calibration'], final_model=False)
@@ -110,7 +58,7 @@ def make_model(given_name, datasets, model_type, model_params, post_params, logg
 
     # If just regular train/test split has been applied
     else:
-        clf = trainEBM(X_train, y_train, model_params, model_type, logging)
+        clf = globals()[model_name].trainModel(X_train, y_train, model_params, model_type, logging)
 
         # discrete prediction
         y_test_pred = clf.predict(X_test)
@@ -127,7 +75,7 @@ def make_model(given_name, datasets, model_type, model_params, post_params, logg
     if post_params['calibration'] != 'false':
         X_all, X_cal, y_all, y_cal = train_test_split(X_all, y_all, test_size=0.2, random_state=123)
 
-    clf = trainEBM(X_all, y_all, model_params, model_type, logging)
+    clf = globals()[model_name].trainModel(X_all, y_all, model_params, model_type, logging)
 
     # Save model in pickled format
     filename = given_name + '/model/ebm_{model_type}.sav'.format(model_type=model_type)
@@ -152,7 +100,7 @@ def make_model(given_name, datasets, model_type, model_params, post_params, logg
             y_all_prob = cal_clf.predict_proba(X_all)[:, 1]
 
 
-
+    # Performance and other post modeling plots
     if model_type == 'classification':
         # Threshold dependant
         plotConfusionMatrixSave(given_name, y_all, y_all_pred, data_type='final_train', logging=logging)
@@ -217,7 +165,7 @@ def make_model(given_name, datasets, model_type, model_params, post_params, logg
         print('Adjusted R2: {adjustedR2}'.format(adjustedR2=adjustedR2))
         logging.info('Adjusted R2: {adjustedR2}'.format(adjustedR2=adjustedR2))
 
-    # Plot explanation of feature importances
-    featureExplanationSave(clf, given_name + '/feature_importance', post_params['file_type'], logging)
+    # Post modeling plots, specific per model but includes feature importance among others
+    globals()[model_name].postModelPlots(clf, given_name + '/feature_importance', post_params['file_type'], logging)
 
     return clf
